Password-Manager/backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
mysql-connector-python==8.0.32
bcrypt==4.0.1
PyJWT==2.6.0
python-dotenv==1.0.0
